backend/requirements.txt
fastapi
uvicorn
pillow
pandas
scikit-learn
joblib
python-multipart
